feat(admin): CategoryController의 역할을 AdminController로 이동
객체지향적이지 못하다면 다시 수정해야 함
