+++ src/main/java/com/spring/springboard/domain/admin/controller/AdminController.java
@@ -1,13 +1,19 @@
 package com.spring.springboard.domain.admin.controller;
 
 import com.spring.springboard.domain.admin.service.AdminService;
+import com.spring.springboard.domain.category.dto.request.CategoryRequest;
+import com.spring.springboard.domain.category.entity.Category;
+import com.spring.springboard.domain.category.service.CategoryService;
 import com.spring.springboard.domain.user.dto.response.UserResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
+
+import java.util.List;
 
 @RestController
 @RequiredArgsConstructor
@@ -15,6 +21,7 @@ import org.springframework.web.bind.annotation.*;
 @RequestMapping("/admin")
 public class AdminController {
     private final AdminService adminService;
+    private final CategoryService categoryService;
 
     @GetMapping("/users")
     public ResponseEntity<Page<UserResponse>> getAllUsers(Pageable pageable) {
@@ -25,6 +32,30 @@ public class AdminController {
     @PutMapping("/users/restore/{userId}")
     public ResponseEntity<Void> restoreUser(@PathVariable Long userId) {
         adminService.restoreUser(userId);
+        return ResponseEntity.noContent().build();
+    }
+
+    @GetMapping("/categories")
+    public ResponseEntity<List<Category>> getAllCategories() {
+        List<Category> categories = categoryService.getAllCategories();
+        return ResponseEntity.ok(categories);
+    }
+
+    @PostMapping("/categories")
+    public ResponseEntity<Category> createCategory(@RequestBody CategoryRequest request) {
+        Category category = categoryService.createCategory(request.getName(), request.getDescription());
+        return ResponseEntity.status(HttpStatus.CREATED).body(category);
+    }
+
+    @PutMapping("/categories/{categoryId}")
+    public ResponseEntity<Category> updateCategory(@PathVariable Long categoryId, @RequestBody CategoryRequest request) {
+        Category category = categoryService.updateCategory(categoryId, request.getName(), request.getDescription());
+        return ResponseEntity.ok(category);
+    }
+
+    @DeleteMapping("/categories/{categoryId}")
+    public ResponseEntity<Void> deleteCategory(@PathVariable Long categoryId) {
+        categoryService.deleteCategory(categoryId);
         return ResponseEntity.noContent().build();
     }
 }
