--- src/main/java/com/spring/springboard/domain/category/controller/CategoryController.java
@@ -1,42 +1,0 @@
-package com.spring.springboard.domain.category.controller;
-
-import com.spring.springboard.domain.category.dto.request.CategoryRequest;
-import com.spring.springboard.domain.category.entity.Category;
-import com.spring.springboard.domain.category.service.CategoryService;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-
-import java.util.List;
-
-@RestController
-@RequiredArgsConstructor
-@RequestMapping("/admin/categories")
-public class CategoryController {
-    private final CategoryService categoryService;
-
-    @GetMapping
-    public ResponseEntity<List<Category>> getAllCategories() {
-        List<Category> categories = categoryService.getAllCategories();
-        return ResponseEntity.ok(categories);
-    }
-
-    @PostMapping
-    public ResponseEntity<Category> createCategory(@RequestBody CategoryRequest request) {
-        Category category = categoryService.createCategory(request.getName(), request.getDescription());
-        return ResponseEntity.status(HttpStatus.CREATED).body(category);
-    }
-
-    @PutMapping("/{categoryId}")
-    public ResponseEntity<Category> updateCategory(@PathVariable Long categoryId, @RequestBody CategoryRequest request) {
-        Category category = categoryService.updateCategory(categoryId, request.getName(), request.getDescription());
-        return ResponseEntity.ok(category);
-    }
-
-    @DeleteMapping("/{categoryId}")
-    public ResponseEntity<Void> deleteCategory(@PathVariable Long categoryId) {
-        categoryService.deleteCategory(categoryId);
-        return ResponseEntity.noContent().build();
-    }
-}
